Add missing file, and move Annotation and Comment to a separate source file.
