+++ py8dis/format.py
@@ -1,0 +1,8 @@
+from enum import auto, Enum
+
+class Format(Enum):
+    DEFAULT = auto()
+    HEX = auto()
+    DECIMAL = auto()
+    BINARY = auto()
+    CHAR = auto()
